fixed list and iterable tests
